migrate: change sidebar to topics
